rename plugin-loader to extension
