+++ Composer/package.json
@@ -20,11 +20,11 @@
     "packages/adaptive-form",
     "packages/client",
     "packages/electron-server",
+    "packages/extension",
     "packages/extension-client",
     "packages/intellisense",
     "packages/lib",
     "packages/lib/*",
-    "packages/plugin-loader",
     "packages/server",
     "packages/test-utils",
     "packages/tools",
@@ -40,7 +40,7 @@
     "build:test": "yarn workspace @bfc/test-utils build",
     "build:lib": "yarn workspace @bfc/libs build:all",
     "build:electron": "yarn workspace @bfc/electron-server build",
-    "build:extensions": "wsrun -lt -p @bfc/plugin-loader @bfc/intellisense @bfc/extension-client @bfc/adaptive-form @bfc/adaptive-flow @bfc/ui-plugin-*  -c build",
+    "build:extensions": "wsrun -lt -p @bfc/extension @bfc/intellisense @bfc/extension-client @bfc/adaptive-form @bfc/adaptive-flow @bfc/ui-plugin-*  -c build",
     "build:server": "yarn workspace @bfc/server build",
     "build:client": "yarn workspace @bfc/client build",
     "build:tools": "yarn workspace @bfc/tools build:all",
