+++ Composer/packages/extension/src/loader/composerPluginRegistration.ts
