+++ Composer/packages/extension/src/loader/index.ts
