+++ Composer/packages/extension/src/loader/pluginLoader.ts
