+++ Composer/packages/extension/src/manager/index.ts
