+++ Composer/packages/extension/src/storage/extensionManifestStore.ts
