+++ Composer/packages/extension/src/types/extension.ts
