+++ Composer/packages/server/package.json
@@ -59,7 +59,7 @@
     "@bfc/intellisense-languageserver": "*",
     "@bfc/lg-languageserver": "*",
     "@bfc/lu-languageserver": "*",
-    "@bfc/plugin-loader": "*",
+    "@bfc/extension": "*",
     "@bfc/shared": "*",
     "@microsoft/bf-dispatcher": "^4.10.0-preview.141651",
     "@microsoft/bf-lu": "^4.10.0-dev.20200808.5a7c973",
