+++ Composer/packages/server/src/controllers/storage.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 import { Request, Response } from 'express';
-import { PluginLoader } from '@bfc/plugin-loader';
+import { PluginLoader } from '@bfc/extension';
 
 import StorageService from '../services/storage';
 import { Path } from '../utility/path';
