+++ Composer/packages/server/src/server.ts
@@ -17,7 +17,7 @@ import { IntellisenseServer } from '@bfc/intellisense-languageserver';
 import { LGServer } from '@bfc/lg-languageserver';
 import { LUServer } from '@bfc/lu-languageserver';
 import chalk from 'chalk';
-import { pluginLoader, PluginManager } from '@bfc/plugin-loader';
+import { pluginLoader, ExtensionManager } from '@bfc/extension';
 
 import { BotProjectService } from './services/project';
 import { getAuthProvider } from './router/auth';
@@ -50,7 +50,7 @@ export async function start(): Promise<number | string> {
   setEnvDefault('COMPOSER_EXTENSION_DATA', path.resolve(__dirname, '../extensions.json'));
   setEnvDefault('COMPOSER_BUILTIN_PLUGINS_DIR', path.resolve(__dirname, '../../../plugins'));
   setEnvDefault('COMPOSER_REMOTE_PLUGINS_DIR', path.resolve(__dirname, '../../../.composer'));
-  await PluginManager.loadBuiltinPlugins();
+  await ExtensionManager.loadBuiltinPlugins();
   // TODO (toanzian / abrown): load 3P plugins
 
   const { login, authorize } = getAuthProvider();
