+++ Composer/packages/server/src/services/storage.ts
@@ -3,7 +3,7 @@
 
 import * as fs from 'fs';
 
-import { UserIdentity } from '@bfc/plugin-loader';
+import { UserIdentity } from '@bfc/extension';
 
 import { Path } from '../utility/path';
 import { StorageConnection, IFileStorage } from '../models/storage/interface';
