+++ Composer/plugins/azurePublish/package.json
@@ -20,7 +20,7 @@
     "@azure/ms-rest-browserauth": "0.1.4",
     "@azure/ms-rest-nodeauth": "3.0.3",
     "@bfc/indexers": "../../packages/lib/indexers",
-    "@bfc/plugin-loader": "../../packages/plugin-loader",
+    "@bfc/extension": "../../packages/extension",
     "@bfc/shared": "../../packages/lib/shared",
     "@microsoft/bf-lu": "^4.10.0-dev.20200728.930f45e",
     "@microsoft/bf-luis-cli": "^4.10.0-dev.20200721.8bb21ac",
