+++ Composer/plugins/azurePublish/src/index.ts
@@ -7,7 +7,7 @@ import { v4 as uuid } from 'uuid';
 import md5 from 'md5';
 import { copy, rmdir, emptyDir, readJson, pathExists, writeJson, mkdirSync, writeFileSync } from 'fs-extra';
 import { IBotProject } from '@bfc/shared';
-import { JSONSchema7 } from '@bfc/plugin-loader';
+import { JSONSchema7 } from '@bfc/extension';
 import { Debugger } from 'debug';
 
 import { mergeDeep } from './mergeDeep';
