+++ Composer/plugins/mockRemotePublish/src/schema.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { JSONSchema7 } from '@bfc/plugin-loader';
+import { JSONSchema7 } from '@bfc/extension';
 
 const schema: JSONSchema7 = {
   type: 'object',
